ssa-backend/requirements.txt
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.1.1
pipenv==2023.11.15
platformdirs==4.0.0
pycparser==2.21
PyMySQL==1.1.0
PyYAML==6.0.1
setuptools==69.0.2
SQLAlchemy==2.0.23
typing_extensions==4.9.0
virtualenv==20.24.7
Werkzeug==3.0.1
